+++ CHANGELOG.md
@@ -8,6 +8,30 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.0.0-pre.1] - 2024-03-15
+
+### Added
+
+- Added support for the [ARPlane.classifications](xref:UnityEngine.XR.ARFoundation.ARPlane.classifications) data type introduced in AR Foundation 6.0, including new classification values and support for planes with multiple classifications.
+- Added an API for the tracking of 3D bounding boxes. Refer to [Bounding boxes](xref:meta-openxr-bounding-boxes) for more information.
+- Added support for the [Asynchronous TryAddAnchor API](xref:arfoundation-whats-new#asynchronous-anchor-creation) introduced in AR Foundation 6.
+- Added [XRResultStatusExtensions](xref:UnityEngine.XR.OpenXR.Features.Meta.XRResultStatusExtensions), extending the [XRResultStatus](xref:UnityEngine.XR.ARSubsystems.XRResultStatus) type to support conversion to [XrResult](xref:UnityEngine.XR.OpenXR.NativeTypes.XrResult).
+- Added support for `NotAxisAligned` plane detection mode introduced in AR Foundation 6.0.
+- Added support for the [Meshing](xref:meta-openxr-meshing) subsystem.
+
+### Changed
+
+- Changed the AR Foundation dependency version to 6.0.0-pre.7.
+- Changed the minimum version of Unity to 2023.3, as required by AR Foundation 6.0.0-pre.7.
+- Changed the behavior of `MetaOpenXRSessionSubsystem.sessionId` to now return a non-empty, unique Guid value per Meta OpenXR session. You can access the session id using `XRSessionSubsystem.sessionId`.
+- Changed the way the Android Manifest is written to override the `IAndroidManifestRequirementProvider` interface so that direct XML manipulation is not necessary.
+- Changed the names of the Meta Quest features in the Editor to remove the "AR" prefix.
+
+### Fixed
+
+- Fixed an issue where the subsystem descriptor for `MetaOpenXRPlaneSubsystem` incorrectly stated that horizontal plane detection, vertical plane detection, and classification were unsupported. These features have all been supported since the 1.0 release, and the subsystem descriptor now returns the correct values.
+- Fixed a very rare issue where plane detection could possibly fail to update planes due to an unhandled error.
+
 ## [1.0.1] - 2023-11-02
 
 ### Changed